Finalised initial ST4 updates
Updated init of blackd to be correctly async - running a command async seems to not work as expected.
Reduced time to test if port number is free and if black is running.
Unified logger logic and added ability to log to file.
Removed checker and related logic as ST4 as on_exit event.